chore(identity): split extensions files
