+++ services/Huntask.Identity.Service/Application/Models/Result.cs
@@ -1,5 +1,4 @@
 using System.Net;
-using Huntask.Identity.Service.Presentation.Models;
 
 namespace Huntask.Identity.Service.Application.Models;
