+++ services/Huntask.Identity.Service/Presentation/Constants.cs
@@ -1,0 +1,6 @@
+namespace Huntask.Identity.Service.Presentation;
+
+public static class Constants
+{
+  public const string CorrelationIdHeaderName = "X-Correlation-Id";
+};
