+++ services/Huntask.Identity.Service/Presentation/Endpoints/HealthcheckEndpoint.cs
@@ -1,0 +1,14 @@
+namespace Huntask.Identity.Service.Presentation.Endpoints;
+
+public static class HealthcheckEndpoint
+{
+  public static WebApplication RegisterHealthcheckEndpoint(this WebApplication app)
+  {
+    app
+      .MapGet("api/healthcheck", () => Results.Ok("Identity Service is running"))
+      .WithName("Healthcheck")
+      .WithTags("Healthcheck");
+
+    return app;
+  }
+}
