+++ services/Huntask.Identity.Service/Presentation/Extensions/ServiceCollectionExtensions.Decorations.cs
@@ -1,0 +1,46 @@
+using CorrelationId.DependencyInjection;
+using Huntask.Identity.Service.Presentation.Swagger;
+
+namespace Huntask.Identity.Service.Presentation.Extensions;
+
+public static class ServiceCollectionDecorationsExtensions
+{
+  public static IServiceCollection AddDecorations(this IServiceCollection services)
+  {
+    services
+      .AddCorrelationId()
+      .AddCache()
+      .AddSwagger();
+
+    return services;
+  }
+
+  private static IServiceCollection AddCorrelationId(this IServiceCollection services)
+  {
+    services.AddDefaultCorrelationId(options =>
+    {
+      options.IncludeInResponse = true;
+      options.RequestHeader = Constants.CorrelationIdHeaderName;
+      options.UpdateTraceIdentifier = true;
+    });
+
+    return services;
+  }
+
+  private static IServiceCollection AddCache(this IServiceCollection services)
+  {
+    services.AddOutputCache();
+    services.AddResponseCaching();
+
+    return services;
+  }
+
+  private static IServiceCollection AddSwagger(this IServiceCollection services)
+  {
+    services.AddEndpointsApiExplorer();
+    services.AddSwaggerGen();
+    services.ConfigureOptions<ConfigureSwaggerOptions>();
+
+    return services;
+  }
+}
