+++ services/Huntask.Identity.Service/Program.cs
@@ -3,13 +3,18 @@ using Huntask.Identity.Service.Presentation.Extensions;
 using Huntask.Identity.Service.Presentation.Endpoints;
 
 var builder = WebApplication.CreateBuilder(args);
-builder.Services.RegisterServices(builder.Configuration);
-builder.Services.RegisterAspNetServices(builder);
+builder.Services
+  .AddInfrastructureServices(builder.Configuration)
+  .AddPresentationServices()
+  .AddDecorations();
 
 var app = builder
   .ConfigureBuilder()
   .Build();
-app.ConfigureWebApplication();
-app.RegisterHealthcheckEndpoint();
+
+app
+  .ConfigurePresentation()
+  .ConfigureDecorations()
+  .RegisterHealthcheckEndpoint();
 
 app.Run();
